requirements.txt
fastapi
fastapi[standard]
sqlalchemy
python-dotenv